Moved to cmd
